requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
python-dotenv
sentence-transformers
torch